Moved ClassNameFilter to com.hz.nio.serialization
