+++ hazelcast/src/main/java/com/hazelcast/nio/serialization/ClassNameFilter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.nio;
+package com.hazelcast.nio.serialization;
 
 /**
  * Allows to intercept class resolution during deserialization based on classname. It's used as a validation mechanism in
