+++ hazelcast/src/test/java/com/hazelcast/nio/SerializationClassNameFilterTest.java
@@ -17,6 +17,7 @@
 package com.hazelcast.nio;
 
 import com.hazelcast.config.JavaSerializationFilterConfig;
+import com.hazelcast.nio.serialization.SerializationClassNameFilter;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
 import org.junit.Test;
